Added Spring Security and Added Cart And Orders Operations

+++ src/main/java/com/developerteam/techzone/business/concreates/UserAdressManager.java
@@ -1,25 +1,34 @@
 package com.developerteam.techzone.business.concreates;
 
+import com.developerteam.techzone.business.abstracts.IAuthService;
 import com.developerteam.techzone.business.abstracts.IUserAdressService;
 import com.developerteam.techzone.dataAccess.abstracts.IUserAdressRepository;
+import com.developerteam.techzone.dataAccess.abstracts.IUserRepository;
+import com.developerteam.techzone.entities.concreates.User;
 import com.developerteam.techzone.entities.concreates.UserAdress;
+import com.developerteam.techzone.entities.dto.DtoUserAdress;
+import com.developerteam.techzone.entities.dto.DtoUserAdressIU;
+import org.springframework.beans.BeanUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
 
 @Service
 public class UserAdressManager  implements IUserAdressService {
 
+    @Autowired
     private IUserAdressRepository userAdressRepository;
 
     @Autowired
-    public UserAdressManager(IUserAdressRepository userAdressRepository) {
-        this.userAdressRepository = userAdressRepository;
-    }
+    private IUserRepository userRepository;
 
+    @Autowired
+    private IAuthService authService;
 
+    //For Admin Admin için yetki kontrolü ekle
     @Override
     public List<UserAdress> getAll() {
         return userAdressRepository.findAll();
@@ -27,6 +36,7 @@ public class UserAdressManager  implements IUserAdressService {
 
     @Override
     public UserAdress getById(int id) {
+        Optional <User> optionalUser = authService.getAuthenticatedUser();
         return userAdressRepository.findById(id).orElse(null);
     }
 
@@ -35,29 +45,85 @@ public class UserAdressManager  implements IUserAdressService {
         return userAdressRepository.findByUserId(userId);
     }
 
-    @Override
-    public UserAdress add(UserAdress userAdress) {
-        return userAdressRepository.save(userAdress);
+
+    /****
+     *
+     * **/
+
+
+    //For Customer
+
+    private boolean currentAdressHasUser(User user, UserAdress userAdress) {
+        return user.getId()== userAdress.getUser().getId();
     }
 
     @Override
-    public UserAdress update(int id, UserAdress userAdress) {
-        Optional <UserAdress> existingUserAdress = userAdressRepository.findById(id);
-        if (existingUserAdress.isPresent()){
-            UserAdress updateUserAdress = existingUserAdress.get();
-            updateUserAdress.setCountry(userAdress.getCountry());
-            updateUserAdress.setCity(userAdress.getCity());
-            updateUserAdress.setDistrict(userAdress.getDistrict());
-            updateUserAdress.setPostCode(userAdress.getPostCode());
-            updateUserAdress.setAdress(userAdress.getAdress());
+    public List<DtoUserAdress> getOwnAdressByUserId() {
+        Optional <User> optionalUser = authService.getAuthenticatedUser();
+        List<UserAdress> dbUserAdressList = userAdressRepository.findByUserId(optionalUser.get().getId());
+        List<DtoUserAdress> dtoUserAdressList = new ArrayList<>();
 
-            return userAdressRepository.save(updateUserAdress);
+        for (UserAdress userAdress : dbUserAdressList) {
+            DtoUserAdress dtoUserAdress = new DtoUserAdress();
+            BeanUtils.copyProperties(userAdress, dtoUserAdress);
+            dtoUserAdressList.add(dtoUserAdress);
         }
-        return null;
+
+        return dtoUserAdressList;
+    }
+
+    @Override
+    public DtoUserAdress add(DtoUserAdressIU dtoUserAdressIU) {
+        Optional<User> optionalUser = authService.getAuthenticatedUser();
+        UserAdress userAdress = new UserAdress();
+        BeanUtils.copyProperties(dtoUserAdressIU, userAdress);
+        userAdress.setUser(optionalUser.get());
+        UserAdress userAdressSaved = userAdressRepository.save(userAdress);
+        DtoUserAdress response = new DtoUserAdress();
+        BeanUtils.copyProperties(userAdressSaved, response);
+        return response;
+    }
+
+    @Override
+    public DtoUserAdress update(int id, DtoUserAdressIU dtoUserAdressIU) {
+        Optional <UserAdress> existingUserAdress = userAdressRepository.findById(id);
+        Optional<User> optionalUser = authService.getAuthenticatedUser();
+
+        if(existingUserAdress.isEmpty()){
+            return null;
+        }
+
+        if (optionalUser.isEmpty()) {
+            return null;
+        }
+        if (!currentAdressHasUser(optionalUser.get(), existingUserAdress.get())) {
+            return null;
+        }
+
+        UserAdress userAdress = new UserAdress();
+        BeanUtils.copyProperties(dtoUserAdressIU, userAdress);
+        userAdress.setUser(optionalUser.get());
+        UserAdress userAdressSaved = userAdressRepository.save(userAdress);
+        DtoUserAdress dtoUserAdress = new DtoUserAdress();
+        BeanUtils.copyProperties(userAdressSaved, dtoUserAdress);
+        return dtoUserAdress;
+
     }
 
     @Override
     public void delete(int id) {
+        Optional<UserAdress> existingUserAdress = userAdressRepository.findById(id);
+        if (existingUserAdress.isEmpty()) {
+            return;
+        }
+        Optional<User> optionalUser = authService.getAuthenticatedUser();
+        if (optionalUser.isEmpty()){
+            return;
+        }
+        UserAdress userAdress = existingUserAdress.get();
+        if (!currentAdressHasUser(optionalUser.get(), userAdress)) {
+            return;
+        }
         userAdressRepository.deleteById(id);
 
     }
